Make sure clicking on GameCard's is only possible within Schedule
The previously added functionality also led to the game cards on
the bets page being clickable. This is not what we want,
so I refactored the GameCard to fix this behaviour.
In general, the GameCard component should be more flexible
and it might need some bigger refactoring when introducing
recent games functionality on the home page.