Fix broken overlay/save panel margins
After removing the margin on the parent element, the panel
and overlay do not need the negative 20px margins anymore,
which previously were required in order to have it cover the
whole viewport width. Since removing the margin on the parent
these properties led to those elements being too wide.