Solved sticky add post button with vanilla CSS
Next time google first...
This is much faster because it does not need the expensive
scroll event handler.
Source: https://gedd.ski/post/position-sticky/

This does not work on IE & Edge though, which is in this case
OK, but we might consider polyfilling it (requires JS)
for the bets page.