Improve zero comments label and behaviour of button
Since one can comment a Post now directly below the meta info,
it doesn't make sense to show the comment are when clicking
on the comment count button. It should be disabled when there
are no comments.
Also change the label so that it does not say "comment",
because that's not what the button does anymore.