Refactor save/error callbacks of GameCardBet
Offer dedicated success and failure callbacks.
On success, return the entire new/updated bet object in the
callback, instead of just the result_bet string (this was done
so far).
Slightly adjust the handler in GameBetsTab to deal with the new
callbacks.
Finish implementing the CurrentGameCard callbacks, which can
now update the CurrentGames state.bets array with entire
updated bets properly.