Further fix of adding top level comments vs. replies
Not as easy as it seemed... added comment/replies need to be
propagated up to the Post, which now must keep track of
all comments in its state.
This is a pretty unsatisfying solution and might need later cleanup.
This also produced another bug which is that leaf replies
with hierarchyLevel > 1 are not display initially anymore.