Add scrollable option to StandingsTable
Then all rows are loaded into the DOM, and the scrollable table
content is automatically scrolled to center the users rank.
It is yet to be solved to not show the fading top/bottom
when the user scrolls to the top/bottom of the table, but
I guess we need a JS scrollListener for this and we
then maybe better leave the fading bits out entirely.