Always filter by user when requesting bets
Otherwise also bets from other user will be returned after
the deadline has passed! This would have led to wrong
bets being shown to the user in the current games and schedule
sections (we .find bets based on bettable ID and the first
match is returned).
An exception (and the reason why I noticed this) is when
bets are requested in the StandingsTable in order to show
other users' bets after the deadline has passed.