Remove scroll listener on unmount
With the singleton scroll listener just implemented, this is
a bit dirty, because the Header and ScrollToTopButton use
the same listener. It's okay because we know that both are
present on each page and removing it twice doesn't do any harm,
but probably the singleton pattern for the scroll listener
is not the best solution in the end.