Allow other than the default global scroll listener
Weaken the previously introduced singleton pattern on the
global scroll listener a bit, in that the throttled scroll listener
class is again exported from EventsHelper and additional
listeners may be registered.
Also renamed the "singleton" instance, which can be used by
clients as long as they know how to clean it up, to be more
clear.