Revert sharing the same scroll listener for Header and ScrollToTopButton
The idea of sharing the same listener instance with multiple
callbacks did have uncontrollable side effects.
Header and/or ScrollToTopButton added and removed the handler
without my control due to sharing the same instance and
not being able to control the component lifecycle.
It would have been better performance-wise, but seemingly
cannot be done that easily.