Make Bets tabs also sticky
This was a bit tricky, because Material UI has bad support
for giving individual class names to the tabs wrapper and the
ink bar (which is a sibling of the wrapper). The resulting Sass
is a bit messy, but it works fine for now.

In the Header, we now also add css classes to identify if
a header-offset has to be applied or not, in order
to be able handle the offset for the ugly bets tabs selectors in the
Sass and not in the JS selectors array.

Let's not try to support IE for now, because I expect this to even
get a lot messier when we attempt to achieve the correct behaviour
using stickybits. The previous IE behaviour doesn't break.