Remove number of 'Volltreffer' from rank distinction
We decided not to use any criteria to distinguish between
users with the identical number of points. They will simply
be on the same rank, no matter how many bets they placed or
how many 'Volltreffer' they've achieved.

We also added a short statement about this on the Rules page.