Start implementing pagination for news
Only load up to 5 news initially, load more posts on request.
This requires to only load posts with news_appear = True, in order
to have a correct pagination. This in turn means that drafts
may never be saved with that flag set to False, because they
would then disappear on reload.

This could be improved by separately requesting the potential
draft in a second request. This would also avoid the pagination
to be messed up by additional dirty drafts in the data
(which should never happen in theory).
However, for now this should do, it's only an admin feature
in the end...