Add lazy loading functionality to BigPicture component
It now can be configured to only load the image when it's about
to be in the viewport of the user.
This property is false by default and should be used sparsely,
since it needs to register a scroll and resize listener on each
corresponding image.
This will be useful especially for the Foyer page, where a lot
of the initial loading footprint can be saved by lazily loading
the pictures which are not initially in view.