Fix redirect bug on invalid credentials
Introduce flag to prevent the FetchHelper to check if the
user is still authenticated.
This should not be done during
the log in request itself, so the
user is not generically redirected
to the login page on a 401 response,
but can instead see an appropriate
error message without page reload.