Remove unsafe CWRP for UserDetailsPopover
In order to lazily load the UserDetailsPopover's
contents, make sure to only mount it when
it's triggered instead of mounting all popovers
on page load and loading the details on open.

This way, componentDidMount can safely
be used to fetch the data.
Plus, we avoid unnecessary DOM elements which
might never be seen by the user.