Update @date-io and date-fns dependencies
Why do we still need the explicit date-fns dependency? As I understand
the @date-io abstraction, it should be possible to rely only on that
without the need of adding date-fns as well. However, I don't yet fully
understand how the date-fns imports then need to be adjusted. This needs
further investigation.