Introduce custom LoginError class
In order to satisfy the linter's recommendation to use actual Error
objects in Promise rejections, I'm adding a dedicated LoginError class
which suites the schema used in all AuthService error responses.

This makes the API in all catch clauses that access error details more
concise indeed.

Also fix an error with determining the API base URL for the given
environment.