Refactor post comments to eliminate circular dep
In this commit I'm able to get rid of a circular dependency
between CommentsList and Comment components. This could be solved by
recursively rendering <Comment>'s and their replies. <CommentsList> is
only used as top level component in the Post.

This also makes the components more concise and the code a bit more
readable.

I got rid of the "load more replies" button for nested comments as it is
not really required at this time. There are rarely so many nested
comments so far that this would be needed.