Sanitize HTML post content to avoid XSS
This should always be done when HTML content from the backend should be
rendered using React's "dangeoruslySetInnerHTML" property.

Only admins "should" be able to add rich text Posts and the
input is therefore controllable, hence XSS should not be an issue here.
Still, it can't be wrong to be double-safe and sanitize the content
using the small DOMPurify library.