Remove indentation module in Post editor
The indentation does not work in the HTML which is generated by Quill,
because it relies on a CSS class that might not be present (especially
when Quill styles are never lazy-loaded because the editor is never opened).

I don't fully understand how this is meant to be working. I don't do any
further investigation now because we don't really need indentation in
the Post editor. Hence, I just remove it from the Quill toolbar for now.