Add new admin page for entering game results
The page does not have any meaningful content yet.
So far, entering game results is only possible via the default Django
admin backend. This is somewhat annoying, because the page is not
mobile-optimized and a bit tedious to use.
The new page will allow to enter the results for games within the RTG page.