First draft of GameResults component
It yet uses the GameCardBet component in order to display a working game
card, but this is not what we need later on to enter results. Some
refactoring will be necessary which ideally also optimizes the long-ish
GameCardBet component.

Currently all games are fetched for wich the deadline has passed. This
is, however, not correct. It needs to be all games which have been
kicked off. There is a way of filtering this in the URL in Django Rest
Framewor which I still have to figure out.